Add PostionReadable interface
