+++ core/client/fs/src/main/java/alluxio/client/file/cache/TimeBoundPageStore.java
@@ -11,10 +11,10 @@
 
 package alluxio.client.file.cache;
 
-import alluxio.client.file.cache.store.PageReadTargetBuffer;
 import alluxio.client.file.cache.store.PageStoreOptions;
 import alluxio.exception.PageNotFoundException;
 import alluxio.exception.status.ResourceExhaustedException;
+import alluxio.file.ReadTargetBuffer;
 import alluxio.metrics.MetricKey;
 import alluxio.metrics.MetricsSystem;
 
@@ -88,7 +88,7 @@ public class TimeBoundPageStore implements PageStore {
   }
 
   @Override
-  public int get(PageId pageId, int pageOffset, int bytesToRead, PageReadTargetBuffer target,
+  public int get(PageId pageId, int pageOffset, int bytesToRead, ReadTargetBuffer target,
       boolean isTemporary) throws IOException, PageNotFoundException {
     Callable<Integer> callable = () ->
         mPageStore.get(pageId, pageOffset, bytesToRead, target, isTemporary);
