+++ core/client/fs/src/main/java/alluxio/client/file/cache/store/RocksPageStore.java
@@ -15,6 +15,7 @@ import alluxio.Constants;
 import alluxio.client.file.cache.PageId;
 import alluxio.client.file.cache.PageStore;
 import alluxio.exception.PageNotFoundException;
+import alluxio.file.ReadTargetBuffer;
 import alluxio.proto.client.Cache;
 
 import com.google.common.base.Preconditions;
@@ -177,7 +178,7 @@ public class RocksPageStore implements PageStore {
   }
 
   @Override
-  public int get(PageId pageId, int pageOffset, int bytesToRead, PageReadTargetBuffer target,
+  public int get(PageId pageId, int pageOffset, int bytesToRead, ReadTargetBuffer target,
       boolean isTemporary) throws IOException, PageNotFoundException {
     Preconditions.checkArgument(pageOffset >= 0, "page offset should be non-negative");
     try {
