+++ core/client/fs/src/test/java/alluxio/client/file/cache/TimeBoundPageStoreTest.java
@@ -20,12 +20,12 @@ import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 import alluxio.Constants;
-import alluxio.client.file.cache.store.ByteArrayTargetBuffer;
 import alluxio.client.file.cache.store.PageStoreOptions;
 import alluxio.conf.Configuration;
 import alluxio.conf.InstancedConfiguration;
 import alluxio.conf.PropertyKey;
 import alluxio.exception.PageNotFoundException;
+import alluxio.file.ByteArrayTargetBuffer;
 import alluxio.util.io.BufferUtils;
 
 import org.junit.Before;
