+++ core/client/fs/src/test/java/alluxio/client/file/cache/store/MemoryPageStoreTest.java
@@ -16,6 +16,7 @@ import static org.junit.Assert.assertEquals;
 
 import alluxio.client.file.cache.PageId;
 import alluxio.client.file.cache.PageStore;
+import alluxio.file.ByteArrayTargetBuffer;
 
 import org.junit.Before;
 import org.junit.Test;
