+++ core/server/worker/src/test/java/alluxio/worker/page/PagedBlockWriterTest.java
@@ -25,12 +25,12 @@ import alluxio.client.file.cache.PageMetaStore;
 import alluxio.client.file.cache.PageStore;
 import alluxio.client.file.cache.evictor.CacheEvictor;
 import alluxio.client.file.cache.evictor.FIFOCacheEvictor;
-import alluxio.client.file.cache.store.ByteArrayTargetBuffer;
 import alluxio.client.file.cache.store.LocalPageStoreDir;
 import alluxio.client.file.cache.store.PageStoreOptions;
 import alluxio.conf.Configuration;
 import alluxio.conf.InstancedConfiguration;
 import alluxio.conf.PropertyKey;
+import alluxio.file.ByteArrayTargetBuffer;
 import alluxio.util.CommonUtils;
 import alluxio.util.WaitForOptions;
 import alluxio.util.io.BufferUtils;
